Fixed a bug in core::quaternion::slerp which gave unnormalized quaternions when falling back to linear interpolation. Optimized slerp to only use 2 trig functions
Added a flerp function which is an approximation of slerp after normalization, together with an optimized flerp when the interpolant is constant across many calls.

Optimized the boning code for skinned meshes, for a stupid reason it was binary searching for a keyframe at a given timestamp for every bone separately, even though all bones share keyframes.